# feat: implement comprehensive performance benchmark suite