--- a/.golangci.yml
+++ b/.golangci.yml
@@ -1,5 +1,3 @@
-version: "2"
-
 run:
   timeout: 5m
   modules-download-mode: readonly