--- a/.golangci.yml
+++ b/.golangci.yml
@@ -1,12 +1,12 @@
-run:
-  timeout: 5m
-  modules-download-mode: readonly
-
+---
 linters:
   enable:
-    # Critical linters that are available in v2 and enabled by default
     - errcheck
     - govet
     - ineffassign
     - staticcheck
     - unused
+  disable-all: false
+
+run:
+  timeout: 5m