--- a/api/api_test.go
+++ b/api/api_test.go
@@ -566,4 +566,3 @@ func TestFinalizePendingClasses(t *testing.T) {
 		assert.True(t, builder.finalized)
 	})
 }
-