--- a/internal/application/service.go
+++ b/internal/application/service.go
@@ -19,16 +19,16 @@ import (
 // TrafficControlService is the main application service that coordinates
 // between the API layer, CQRS handlers, and infrastructure
 type TrafficControlService struct {
-	eventStore         eventstore.EventStoreWithContext
-	netlinkAdapter     netlink.Adapter
-	commandBus         *CommandBus
-	genericCommandBus  *GenericCommandBus
-	queryBus           *QueryBus
-	eventBus           *EventBus
-	projectionManager  *projections.Manager
-	readModelStore     projections.ReadModelStore
-	statisticsService  *StatisticsService
-	logger             logging.Logger
+	eventStore        eventstore.EventStoreWithContext
+	netlinkAdapter    netlink.Adapter
+	commandBus        *CommandBus
+	genericCommandBus *GenericCommandBus
+	queryBus          *QueryBus
+	eventBus          *EventBus
+	projectionManager *projections.Manager
+	readModelStore    projections.ReadModelStore
+	statisticsService *StatisticsService
+	logger            logging.Logger
 }
 
 // NewTrafficControlService creates a new traffic control service