--- a/internal/application/service.go
+++ b/internal/application/service.go
@@ -92,12 +92,12 @@ func (s *TrafficControlService) registerHandlers() {
 	s.commandBus.Register("CreateFilter", chandlers.NewCreateFilterHandler(s.eventStore))
 
 	// Register new type-safe generic command handlers
-	RegisterHandlerFor(s.genericCommandBus, chandlers.NewGenericCreateHTBQdiscHandler(s.eventStore))
-	RegisterHandlerFor(s.genericCommandBus, chandlers.NewGenericCreateHTBClassHandler(s.eventStore))
-	RegisterHandlerFor(s.genericCommandBus, chandlers.NewGenericCreateFilterHandler(s.eventStore))
-	RegisterHandlerFor(s.genericCommandBus, chandlers.NewGenericCreateTBFQdiscHandler(s.eventStore))
-	RegisterHandlerFor(s.genericCommandBus, chandlers.NewGenericCreatePRIOQdiscHandler(s.eventStore))
-	RegisterHandlerFor(s.genericCommandBus, chandlers.NewGenericCreateFQCODELQdiscHandler(s.eventStore))
+	RegisterHandlerFor[*models.CreateHTBQdiscCommand](s.genericCommandBus, chandlers.NewGenericCreateHTBQdiscHandler(s.eventStore))
+	RegisterHandlerFor[*models.CreateHTBClassCommand](s.genericCommandBus, chandlers.NewGenericCreateHTBClassHandler(s.eventStore))
+	RegisterHandlerFor[*models.CreateFilterCommand](s.genericCommandBus, chandlers.NewGenericCreateFilterHandler(s.eventStore))
+	RegisterHandlerFor[*models.CreateTBFQdiscCommand](s.genericCommandBus, chandlers.NewGenericCreateTBFQdiscHandler(s.eventStore))
+	RegisterHandlerFor[*models.CreatePRIOQdiscCommand](s.genericCommandBus, chandlers.NewGenericCreatePRIOQdiscHandler(s.eventStore))
+	RegisterHandlerFor[*models.CreateFQCODELQdiscCommand](s.genericCommandBus, chandlers.NewGenericCreateFQCODELQdiscHandler(s.eventStore))
 
 	// Register query handlers with read model support
 	// Note: Query handlers temporarily disabled due to interface compatibility issues