… CLAUDE.md
… README.md
… docs/benchmarks.md
… internal/application/command_bus_generics_test.go
… internal/commands/handlers/generic_other_handlers.go
… internal/infrastructure/eventstore/benchmark_test.go
… pkg/tc/bandwidth_test.go
… pkg/tc/handle_test.go